agentic-repos-program/3-Litellm_openai_agent/1_LLM_Compare_by_litellm/LLM_Insight/pyproject.toml
[project]
name = "llm-insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.72.0",
    "python-dotenv>=1.1.0",
]
